requirements.txt
# PDF Password Remover Requirements
# Core dependencies
PyPDF2>=3.0.0
PyCryptodome>=3.15.0

# GUI dependencies (for tkinter - usually included with Python)
# tkinter should be available in standard Python installations

# Development dependencies
# Uncomment below for development
# pytest>=7.0.0
# pytest-cov>=4.0.0
# pyinstaller>=5.0.0

# Optional: For enhanced drag-and-drop in GUI
# tkinterdnd2>=0.3.0